requirements.txt
django==3.2
wemake-python-styleguide==0.15.3
bandit==1.7.2
pre-commit==2.15.0
python-dotenv==1.0.0
django-split-settings==1.2.0
psycopg2==2.9.7
pytest==6.2.5
